refactor!: move `RequestReactionType` to model
