+++ twilight-model/Cargo.toml
@@ -14,6 +14,7 @@ version = "0.15.0"
 
 [dependencies]
 bitflags = { default-features = false, version = "1" }
+percent-encoding = { default-features = false, version = "2" }
 serde = { default-features = false, features = ["derive", "std"], version = "1.0.103" }
 serde-value = { default-features = false, version = "0.7" }
 serde_repr = { default-features = false, version = "0.1.5" }
