+++ twilight-model/src/http/mod.rs
@@ -3,3 +3,7 @@
 pub mod attachment;
 pub mod interaction;
 pub mod permission_overwrite;
+
+mod reaction_type;
+
+pub use self::reaction_type::RequestReactionType;
